+++ spec-main/fixtures/module/preload-remote-function.js
